fixed the internal type for the push token
the token only used a uint8_t which would silently overflow and cause
very difficult to debug behavior.
updated the docs to make the maximum value and the behavior clearer.

+++ tlc5947/tlc5947.c
@@ -117,7 +117,7 @@ typedef struct _token_t{
         struct{                                        }forever;
         struct{uint16_t new_pp;                        }jump;
         struct{                                        }mark;
-        struct{int8_t value;                           }push;
+        struct{int16_t value;                          }push;
         struct{                                        }pop;
     };
 }token_t;
